backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
celery>=5.3
django-celery-beat>=2.5
requests>=2.31
gunicorn>=21.2
